[Python] update python authentication to skip empty token/username, password

+++ samples/client/petstore/python/swagger_client/apis/store_api.py
@@ -78,6 +78,7 @@ class StoreApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/store/inventory'.replace('{format}', 'json')
         method = 'GET'
 
@@ -151,6 +152,7 @@ class StoreApi(object):
             params[key] = val
         del params['kwargs']
 
+
         resource_path = '/store/order'.replace('{format}', 'json')
         method = 'POST'
 
@@ -212,9 +214,6 @@ class StoreApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'order_id' is set
-        if order_id is None:
-            raise ValueError("Missing the required parameter `order_id` when calling `get_order_by_id`")
 
         all_params = ['order_id']
         all_params.append('callback')
@@ -228,6 +227,10 @@ class StoreApi(object):
                 )
             params[key] = val
         del params['kwargs']
+
+        # verify the required parameter 'order_id' is set
+        if ('order_id' not in params) or (params['order_id'] is None):
+            raise ValueError("Missing the required parameter `order_id` when calling `get_order_by_id`")
 
         resource_path = '/store/order/{orderId}'.replace('{format}', 'json')
         method = 'GET'
@@ -290,9 +293,6 @@ class StoreApi(object):
                  If the method is called asynchronously,
                  returns the request thread.
         """
-        # verify the required parameter 'order_id' is set
-        if order_id is None:
-            raise ValueError("Missing the required parameter `order_id` when calling `delete_order`")
 
         all_params = ['order_id']
         all_params.append('callback')
@@ -306,6 +306,10 @@ class StoreApi(object):
                 )
             params[key] = val
         del params['kwargs']
+
+        # verify the required parameter 'order_id' is set
+        if ('order_id' not in params) or (params['order_id'] is None):
+            raise ValueError("Missing the required parameter `order_id` when calling `delete_order`")
 
         resource_path = '/store/order/{orderId}'.replace('{format}', 'json')
         method = 'DELETE'
